Consolidate main forms
